refactor: move files out of the src folder
