+++ .github/workflows/publish.yml
@@ -1,7 +1,7 @@
 name: Publish Python Package
 
 on:
-  pull_request:
+  push:
     tags:
       - "v*"
 
@@ -22,8 +22,5 @@ jobs:
           pip install hatch
       - name: Build package
         run: hatch build
-      - name: Clear GZ
-        run: |
-          find dist -name "pyopenagi-*.tar.gz" -delete
       - name: Publish package distributions to PyPI
         uses: pypa/gh-action-pypi-publish@release/v1
