+++ pyopenagi/agents/agent_config/MathAgent.json
