+++ pyopenagi/tools/base.py
@@ -1,4 +1,4 @@
-from src.utils.utils import get_from_env
+from pyopenagi.utils.utils import get_from_env
 import json
 
 class BaseTool:
