+++ pyopenagi/tools/online/bing_search.py
@@ -6,7 +6,7 @@ from ..base import BaseTool
 
 # from pydantic import root_validator
 
-from src.utils.utils import get_from_env
+from pyopenagi.utils.utils import get_from_env
 
 class BingSearch(BaseTool):
     """Bing Search Tool, refactored from langchain.
