+++ pyopenagi/tools/online/shazam/song_auto_complete.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, List, Optional
 
 # from pydantic import root_validator
 
-from openagi.src.utils.utils import get_from_env
+from pyopenagi.utils.utils import get_from_env
 
 import requests
