+++ pyopenagi/tools/online/trip_advisor/hotels.py
@@ -2,7 +2,7 @@ from ...base import BaseRapidAPITool
 
 from typing import Any, Dict, List, Optional
 
-from src.utils.utils import get_from_env
+from pyopenagi.utils.utils import get_from_env
 
 import requests
