+++ pyopenagi/tools/online/wikipedia.py
