+++ pyopenagi/tools/online/wolfram_alpha.py
@@ -2,7 +2,7 @@ from typing import Any, Dict, Optional
 
 from ..base import BaseTool
 
-from src.utils.utils import get_from_env
+from pyopenagi.utils.utils import get_from_env
 class WolframAlpha(BaseTool):
     """Wolfram Alpha Tool, refactored from langchain.
