+++ pyopenagi/tools/online/words_api.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, List, Optional
 
 # from pydantic import root_validator
 
-from src.utils.utils import get_from_env
+from pyopenagi.utils.utils import get_from_env
 
 import requests
