+++ tests/test_tools/test_currency_converter.py
@@ -1,7 +1,7 @@
 import os
 import pytest
 
-from src.tools.online.words_api import WordsAPI
+from pyopenagi.tools.online.currency_converter import CurrencyConverterAPI
 from dotenv import load_dotenv, find_dotenv
 
 @pytest.fixture(scope="module")
@@ -9,16 +9,20 @@ def test_rapid_api_key():
     load_dotenv(find_dotenv())
     if "RAPID_API_KEY" not in os.environ or not os.environ["RAPID_API_KEY"]:
         with pytest.raises(ValueError):
-            words_api = WordsAPI()
+            currency_converter_api = CurrencyConverterAPI()
         pytest.skip("Rapid api key is not set.")
+    else:
+        return True
 
 @pytest.mark.usefixtures("test_rapid_api_key")
-def test_words_api():
-    words_api = WordsAPI()
+def test_currency_converter_api():
+    load_dotenv(find_dotenv())
+    currency_converter_api = CurrencyConverterAPI()
     params = {
-        "word": "look",
-        "api_name": "typeOf",
+        "from": "USD",
+        "to": "EUR",
+        "amount": 2
     }
-    result = words_api.run(params=params)
+    result = currency_converter_api.run(params=params)
     print(result)
     assert isinstance(result, str)
