+++ tests/test_tools/test_words_api.py
@@ -1,7 +1,7 @@
 import os
 import pytest
 
-from src.tools.online.currency_converter import CurrencyConverterAPI
+from pyopenagi.tools.online.words_api import WordsAPI
 from dotenv import load_dotenv, find_dotenv
 
 @pytest.fixture(scope="module")
@@ -9,20 +9,16 @@ def test_rapid_api_key():
     load_dotenv(find_dotenv())
     if "RAPID_API_KEY" not in os.environ or not os.environ["RAPID_API_KEY"]:
         with pytest.raises(ValueError):
-            currency_converter_api = CurrencyConverterAPI()
+            words_api = WordsAPI()
         pytest.skip("Rapid api key is not set.")
-    else:
-        return True
 
 @pytest.mark.usefixtures("test_rapid_api_key")
-def test_currency_converter_api():
-    load_dotenv(find_dotenv())
-    currency_converter_api = CurrencyConverterAPI()
+def test_words_api():
+    words_api = WordsAPI()
     params = {
-        "from": "USD",
-        "to": "EUR",
-        "amount": 2
+        "word": "look",
+        "api_name": "typeOf",
     }
-    result = currency_converter_api.run(params=params)
+    result = words_api.run(params=params)
     print(result)
     assert isinstance(result, str)
